+++ src/app/[locale]/(public)/cart/_actions/remove-specific-cart-item.action.ts
@@ -1,24 +1,31 @@
 'use server';
 
+import { getToken } from '@/lib/utils/get-token';
 import { revalidateTag } from 'next/cache';
 
-export async function removeCartItem(itemId: string , productTitle: string) {
-  const token =
-    'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyIjoiNjhlZTA3YWY3ZmVlNjhhNGMyZWJhZmJhIiwicm9sZSI6InVzZXIiLCJpYXQiOjE3NjIyNzU2MTl9.MQWALjbJZjxiOu-SdCj29ZXhPqcpZcYPRVsacqW8Jfc';
+export async function removeCartItem(itemId: string, productTitle: string) {
   try {
+    const jwt = await getToken();
+
+    if (!jwt || !jwt.token) {
+      throw new Error('Unauthorized');
+    }
+
     const response = await fetch(`${process.env.BASE_URL}/cart/${itemId}`, {
       method: 'DELETE',
       headers: {
         'Content-Type': 'application/json',
-        Authorization: `Bearer ${token}`,
+        Authorization: `Bearer ${jwt.token}`,
       },
     });
+
     if (!response.ok) {
-      throw new Error('Failed to clear cart');
+      throw new Error('Failed to remove cart item');
     }
+
     revalidateTag('cart-data');
     return `${productTitle} removed`;
   } catch (error) {
-    console.error('Error clearing cart:', error);
+    console.error('Error removing cart item:', error);
   }
 }
